+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.9.37+22
+
+* Ajuste header ZEstrutura
+
+## 0.9.37+21
+
+* Ajuste header ZEstrutura
+
+## 0.9.37+20
+
+* Ajuste ZSelection
+
 ## 0.9.37+19
 
 * Ajuste ZEstruturaEmpresa